feat: extract vtcode-indexer crate
