+++ vtcode-indexer/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "vtcode-indexer"
+version = "0.30.0"
+edition = "2024"
+authors = ["vinhnx <vinhnx@users.noreply.github.com>"]
+description = "Workspace-friendly code indexer extracted from VTCode"
+license = "MIT"
+publish = false
+
+[dependencies]
+anyhow = "1.0"
+regex = "1.10"
+serde = { version = "1.0", features = ["derive"] }
+
+[dev-dependencies]
+tempfile = "3.0"
+
